templated classes `Steady` and `Transient`
the SteadyCompressorSystem and TransientCompressorSystem classes are
templated and discretizes the network during first run. In theory this
should support AUTO_DIFF although it is not working yet!. These
templates have however been verified with NUMERIC_DIFF. Additional
resolutions are necessary to eliminate instances of `DiscreteNetwork`
class in the `four_compressor_types` demo.